* Renamed PadTill to PadUntill. * Added FillUntil formatting segment, which pads with a specific character up until the message reaches the specified length. * Added version of profiling that generates a markdown table. * Updated README.
